cnn/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
torch==2.1.0
torchvision==0.16.0
triton==2.1.0
typing-extensions==4.5.0
urllib3==2.0.7
werkzeug==3.0.1
wrapt==1.15.0
zipp==3.17.0
